requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-generativeai==0.7.2
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
Pillow==10.3.0
